# Make summary, wordcount etc. more effective