In total, the effect with this branch compared with master:

```
benchmark           old ns/op      new ns/op      delta
BenchmarkHugo-4     3166246627     3038050615     -4.05%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     10557420       10540152       -0.16%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     1621203992     1605716520     -0.96%
```

See: https://github.com/bep/hugo-benchmark

Word counting benchmark:

``` 
BenchmarkTotalWords-4            100000         18795 ns/op           0  B/op           0 allocs/op 
BenchmarkTotalWordsOld-4          30000         46751 ns/op           6400 B/op           1 allocs/op
```

New auto-summary benchmark:

``` 
BenchmarkTestTruncateWordsToWholeSentence-4            300000         4720 ns/op         0 B/op              0 allocs/op
BenchmarkTestTruncateWordsToWholeSentenceOld-4         100000         17699 ns/op        3072 B/op           3 allocs/op
```